Handle scientific notation & dynamic signal length
When scientific notation was encountered in a dbc file, it was
manifesting itself as a physical value of zero in the calculation
of a physical value from a raw value.  Multiplying the parsed gain
value by 1 seems to convert scientifically notated numbers into
a numeric value.

The calc16 attribute instituted a hard-coded bit length of 16 when
calculating the physical value from the raw value.  An attribute
was added called calcdyn which sets the bit length of the signal
depending upon the bit length specified in the dbc file.